Migrate AddFileView and PermissionsDialog
